- Fixed cast exception in Indentification storage. - Implemented disposeOnExit method for Peptizer

+++ trunk/src/main/java/com/compomics/mslims/util/mascot/MascotResultsProcessor.java
@@ -501,7 +501,7 @@ public class MascotResultsProcessor {
                         hm.put(Fragmention.INTENSITY, new Long(new Double(fi.getIntensity()).longValue()));
                         hm.put(Fragmention.IONNAME, fi.getType());
                         hm.put(Fragmention.IONTYPE, new Long(fi.getID()));
-                        hm.put(Fragmention.L_IDENTIFICATIONID, id.getGeneratedKeys()[0]);
+                        hm.put(Fragmention.L_IDENTIFICATIONID, (Long)(id.getGeneratedKeys()[0]));
                         hm.put(Fragmention.L_IONSCORINGID, new Long(fi.getImportance()));
                         hm.put(Fragmention.MASSDELTA, new Double(new BigDecimal(fi.getTheoreticalExperimantalMassError()).setScale(4, BigDecimal.ROUND_HALF_UP).doubleValue()));
                         hm.put(Fragmention.MASSERRORMARGIN, new Double(id.getFragmentMassTolerance()));
